drivers: sunxi-cir: Implement proper probe/release functions
When first adding CIR support, I attempted to implement it in a way that
did not require any changes to Linux. However, that made integration
with the reference-counting driver classes problematic. The CIR driver
needs setup at suspend time (to switch clocks), but could not be left in
reset at resume time. The reference-counting model leaves only three
possibilities, all unsatisfactory:
 - Run both probe and release (breaks Linux if its driver is loaded)
 - Run neither function (breaks Crust if the Linux driver is loaded)
 - Leak references (eventually crashes Crust)

The only satisfactory approach is the one taken already for I2C and RSB:
add suspend/resume hooks to the Linux driver so it expects the device to
be in reset when resuming. Then the usual Crust probe/release function
structure can be used.

Thankfully, this has little impact when running with an unpatched Linux.
The only problem is that no CIR events will be received after resume.
since the driver only access registers from inside the interrupt
handler (which will never run), nothing else breaks.

Signed-off-by: Samuel Holland <samuel@sholland.org>