drivers: cir: Require a non-NULL device in class op
The general rule for for device references being NULL is that the
device_* functions accept NULL arguments, but the individual class
operations do not. cir_poll() was an exception. To satisfy the general
expectation and to make it slightly more obvious that a CIR wakeup
source may not be available, move the NULL check outside, to the caller.

Signed-off-by: Samuel Holland <samuel@sholland.org>