drivers: cir: Minor RC6 decoder size optimizations
Now that the decoder is more reliable, drop the debugging print when a
pulse is too short. It was only marginally useful, as it only logged one
of many possible transitions back to the idle state.

Save an instruction in RC6_DATA_N by only setting the state once. This
control flow arrangment is copied from the NEC decoder.

Signed-off-by: Samuel Holland <samuel@sholland.org>