drivers: css: Remove return value from state setters
Since the generic CSS code ensures that the target power domain and
state are valid before calling any setters, there is no need for the
setters to have their own return value. In fact, the state cache is
already updated regardless of the return value from the setters.

Currently, the setters can only fail when passed an unsupported state.
However, because the setters are not idempotent, this was already
broken: a transition from cluster ON -> RETENTION will fail, but then a
transition from cluster RETENTION -> ON will reset the cluster (oops!).
The solution to this involves providing both the old state and the new
state to the setters, which will be done shortly.

To avoid repeatedly using the same comments to distinguish common code
from overridable default code, move the default functions to their own
file.

Signed-off-by: Samuel Holland <samuel@sholland.org>