drivers: css: Make power state setters idempotent
Currently, the same functions are used to both raise and lower the state
of a power domain. This complicates using intermediate power states like
RETENTION, as it requires checking both the old and new states. The code
for ON -> RETENTION and OFF -> RETENTION transitions is very different.

Handling both directions in the same function is not really useful. An
overall state transition contains either raising or lowering of power
states, but never both. With a few `unreachable()`s sprinkled in the
code, GCC might even be able to realize this, split the cluster/CSS
functions in half, and inline each half at the appropriate call site. Or
even better, we could split the functions at the source level. This
allows them to be inlined wholesale.

Some analysis of the possible state transitions shows that when split
like this, each setter only cares about the old state or the new state,
but never both. Thus there is no multiplication of inputs to check.

The reason for only needing one state input is that the "other" state is
always ON. This makes verifying the idempotence of each setter trivial:
if the setter does nothing when the passed-in state is _also_ ON, then
the setter is idempotent. (All existing setters pass this test.) Making
the setters idempotent removes the need to check for a state change in
the generic code, which removes another layer of conditionals.

Thus the new code is not only more logically sound, but also smaller.

Signed-off-by: Samuel Holland <samuel@sholland.org>