drivers: css: Fix resume when CPU0 is offline
Normally, CPU0 is online, and it is used by Linux as the lead core for
suspend/resume. However, if CPU0 is manually offlined, it will be left
off, and another core will be the lead core. Then during resume, TF-A
will panic when CPU0 starts executing while not in ON or ON_PENDING.
Fix tis by resuming on the lead core, instead of always using CPU0.

Signed-off-by: Samuel Holland <samuel@sholland.org>