drivers: css: Rename the power state definition file
The file was named css_default.c because it contained the default power
state (i.e. the state at boot). However, that is confusing when there is
also a default CSS implementation. So move the power state definition to
a file with a more-specific name. Due to the macro hacks, it will always
be in a file by itself anyway.

Signed-off-by: Samuel Holland <samuel@sholland.org>