drivers: css: Set CSS resets in the proper function
CPU_SYS_RESET is distinct from the cluster resets even on single-cluster
SoCs. It resets the part of the GIC outside the cluster, and controls
access to the CPU_SYS_CFG MMIO range.

For cluster idle, the GIC must continue functioning while the cluster is
powered down. To do this, the CSS reset must be moved out of the cluster
suspend/resume function and put in its proper place.

Signed-off-by: Samuel Holland <samuel@sholland.org>