drivers: css: Implement CPU idle polling
From an SCPI perspective, a CPU in an idle state is equivalent to a CPU
that has been hot-removed. The difference is how the operating system
does interrupt routing. When a CPU is hot-removed, all interrupts are
routed away from it. This is necessary so the OS can still receive them.
However, when a CPU enters an idle state, the interrupt routing is
unchanged.

Thus, while the idle CPU has interrupts disabled locally, those
interrupts can still be observed at the distributor level, as well as
through the GIC wakeup outputs. Allwinner maps the GIC wakeup outputs to
an MMIO register in the R_CPUCFG region. By polling this register (the
exact method will be provided by the platform-specific CSS driver), the
firmware knows when to wake up a CPU out of its idle state. A few things
to note:

  - Since the firmware is single-threaded, worst-case wakeup latency
    happens when one CPU is executing an SCPI command, and all other
    CPUs attempt to wake at the same time.
  - There is no prioritization. If multiple CPUs have interrupts
    pending, they are woken up in numerical order.
  - Because the CSS resume callbacks are idempotent, checking the power
    state in the polling loop is not strictly necessary. However, it
    reduces latency in the common case where the CPU receiving an
    interrupt is already on (and not idle).

Signed-off-by: Samuel Holland <samuel@sholland.org>