drivers: sun50i-a64-css: Hook up CPU idle
The register exposing the GIC IRQ status was found by experimentation.
Because the register is read-only, and TF-A and Linux do not use FIQs,
it is hard to know which bits are implemented. However, this register is
immediately followed by a register with a writable mask of 0x00000f0f.
This suggests the register is similar to the IRQ_FIQ_STATUS/IRQ_FIQ_MASK
pair documented for A83T.

Other possibilities for equivalent registers include IRQ_FIQ_STATUS_CTRL
as documented for R40/A40i/T3 or IRQ_FIQ_STATUS as documented for H6 and
newer. These documented layouts could put the FIQ status bits at either
bits 8-11 or 16-19. However, since the FIQs are unused, it does not
matter for now.

Signed-off-by: Samuel Holland <samuel@sholland.org>