platform: Document additional PRCM registers
For H6 and newer, the BSP references an RTC gate register following the
other PRCM gate registers. This register also exists on older SoCs. The
gate controls register access only, not timekeeping.

SoCs with an integrated analog audio codec control it through a register
in the PRCM. This register is documented in the various user manuals, in
the audio codec section instead of the PRCM section.

Signed-off-by: Samuel Holland <samuel@sholland.org>