drivers: css: Optimize CSS suspend/resume routines
The cpuidle entry/exit latency directly impacts its power efficiency.
So it is important to minimize work done by the CSS callbacks,
especially explicit delays or polling.

There are two main optimizations done here:

First, when powering down domains, do not assert the reset. Only gate
the outputs and remove power. The resets must be asserted during the
power-up procedure anyway, to handle the initial power-up.

The exception to this is CPU 0 on A64/H5, which does not have gates or a
power switch. It must be put into power-on reset (in R_CPUCFG) so when
the cluster is reset (which deasserts CPU 0 core reset and AA64nAA32),
CPU 0 does not start executing the BROM.

Second, take advantage of known register values. When most or all bits
in a register need to be set or cleared, it is more efficient to write
the final value directly. Further, when the CPUCFG MMIO region is reset
along with the cluster, its contents are set to known values. This, for
example, deasserts ACINACTM and L2RSTDISABLE, so those do not need to be
reprogrammed.

Signed-off-by: Samuel Holland <samuel@sholland.org>