drivers: css: Implement cluster idle states
While it is mostly relevant for multi-cluster SoCs where the other
cluster would be sending snoop requests, entering L2 idle can save a
tiny bit of power. It also provides an opportunity for removing clocks
to the cluster.

This is the SCPI state with core = OFF, cluster = RETENTION, css = ON.

Signed-off-by: Samuel Holland <samuel@sholland.org>