debug: Add a feature to record suspend/resume steps
Currently, all MMIO register polling is done without a timeout. Mostly
these are "should never fail" operations, and the firmware is not
prepared to recover from failure anyway. However, the chance for a hang
at any of these locations makes it difficult to know what broke during a
failed resume.

Recording the start of each major step in persistent storage (the RTC)
allows post-mortem analyis by narrowing down the location of the hang.

The firmware major/minor version is included in the step value as a
marker. The step values are stuffed into 16 bits to reduce code size.

Signed-off-by: Samuel Holland <samuel@sholland.org>