debug: Optionally verify DRAM contents during resume
It is possible for suspending and resuming the DRAM controller to
succeed, but still damage the contents of DRAM. In this case, crust's
resume process will complete, and TF-A will successfully resume (since
it runs from SRAM), but Linux will not be able to run.

Checking for DRAM corruption distinguishes this case from other resume
failures, and it allows the firmware to recover (by resetting the SoC)
instead of leaving Linux hanging.

Signed-off-by: Samuel Holland <samuel@sholland.org>