Makefile: Add version.h dependency to all objects
Now that steps.h includes version.h, several files depend on this
generated header. Since there is no dependency declared in the Makefile
for these new users, the build may fail in some environments.

Adding a Makefile rule for each file is not scalable. Extend the build
system to take a list of order-only dependencies to add for all object
files. This allows us to safely include version.h from anywhere. The
compiler's dependency generation adds a non-order-only dependency, so
the appropriate files are still rebuilt after a commit.

Fixes: 7fae4fbf7859 ("debug: Add a feature to record suspend/resume steps")
Signed-off-by: Samuel Holland <samuel@sholland.org>