README.md: Add a note about Linux version compatibility
Crust and Linux must use the same hardware bus to communicate with the
PMIC, or else Linux will fail to resume from sleep, and Crust will fail
to power off the board. Linux recently changed which bus it uses,
breaking the default Crust configuration. The Crust configuration will
be updated in a future commit; at that point, it will be incompatible
with older kernels.

Signed-off-by: Samuel Holland <samuel@sholland.org>