drivers: sun8i-h3-css: Add CSS suspend support
This hooks up the CPU subsystem reset so power can be removed from the
CSS while the system is suspended. The gating and power switch registers
for CPU 0 appear to have no effect. Even when they are gated and off,
deasserting the CSS reset causes CPU 0 to start executing instructions
unless it is individually reset. The change to handle this matches the
code for the A64 CSS.

Signed-off-by: Samuel Holland <samuel@sholland.org>