debug: Avoid clobbering the last step during boot
The step recording functionality was intended to help debug hangs during
suspend/resume. However, for people without serial console access, it is
all but useless. This is because css_set_power_state() is also used for
SMP secondary core bringup and for cpuidle.

cpuidle is no big deal; it is safe and easy to turn off with the kernel
parameter "cpuidle.off=1". Asking users to turn off SMP is a much larger
ask. So let's record STEP_RESUME_CSS outside of css_set_power_state()
in the system resume function. That way it will not be hit during Linux
boot, and users will be able to read the failing step from userspace.

Removing the two other resume steps does not lose much, since the CSS
drivers are already well-tested by cpuidle.

Signed-off-by: Samuel Holland <samuel@sholland.org>