arch: or1k: Rename counter to cycle_counter
Currently, the trusted watchdog is used as a reference for measuring the
cycle counter frequency. However, platforms before A83T do not have a
trusted watchdog. Even on newer platforms, some users want to disable
the watchdog due to interference from Linux. So a generic "system
counter" is needed to replace direct use of the trusted watchdog.

To avoid ambiguity with this new "system counter", be explicit that the
or1k tick timer is the CPU's "cycle counter". This naming would remain
accurate for a future xtensa port of the firmware.

Signed-off-by: Samuel Holland <samuel@sholland.org>