drivers: counter: Add a CNT64 driver
CNT64 is a 64-bit MMIO counter in the R_CPUCFG space. Reading CNT64 is
inefficient, because it requires multiple MMIO accesses. On the other
hand, CNT64 does not depend on any clock gates, and it is located inside
a secure-only MMIO area.

This makes it a good choice for the system counter on A31-generation
SoCs without a memory-mapped timestamp device.

Signed-off-by: Samuel Holland <samuel@sholland.org>