platform: Correct timestamp device base addresses
The timestamp device has two base addresses, so the control and status
registers can be given different permissions. Both should be declared.

H6 was an outlier in that it named the MMIO regions "CNT_R" (presumably
for "read") and "CNT_C" (presumably for "control") in its manual.
Allwinner returned to the "TIMESTAMP" naming in later SoC user manuals.

Signed-off-by: Samuel Holland <samuel@sholland.org>