lib: Only mangle addresses/layouts when big-endian
Currently, the firmware is built with a big-endian toolchain and expects
the AR100 to be in word-invariant mode. However, in the future it will
be possible to build the firmware with a little-endian toolchain and run
the processor in byte-invariant mode. In that case, no address mangling
will be needed.

So mangling should be enabled depending on endianness. It should not be
enabled unconditionally, nor should it depend only on the architecture.

Signed-off-by: Samuel Holland <samuel@sholland.org>