drivers: watchdog: Add a driver for the A31 watchdog
Older SoCs (A31, A23, and A33) do not contain a trusted watchdog, so
another driver is needed. Since R_WDOG is always available, add a driver
for it, and use it by default.

R_WDOG has the advantage of having no clock gate, which prevents Linux
from interfering with its use.

On the other hand, Linux/TF-A/U-Boot use R_WDOG on H6 due to a hardware
bug in the main watchdog. Crust restarting R_WDOG will prevent it from
timing out, even when the timeout is desired. So for now, keep using the
trusted watchdog on H6.

Signed-off-by: Samuel Holland <samuel@sholland.org>