drivers: watchdog: Simplify the system reset API
watchdog_set_timeout() is always used with the intent of triggering a
system reset. Provide a driver operation explicitly for that purpose, to
hide the details of setting the timeout. This allows each driver to
choose the appropriate timeout value, or otherwise reset the system in
the most efficient manner.

Signed-off-by: Samuel Holland <samuel@sholland.org>