platform: Add Kconfig for an A23/A33 platform
This is the first step toward adding support for A23 and A33 boards.
Several drivers must be added before this platform will compile, so
for now, prevent it from being selected.

A23 and A33 are extremely similar SoCs, other than the obvious
difference in the number of CPU cores. Follow the example of A64/H5,
with one platform subdivided into two SOC options.

Since this adds a new option at the beginning of the PLATFORM choice,
it changes the default value. As a result, all of the existing A64/H5
defconfigs must be updated to explicitly list their PLATFORM_A64 choice.

Signed-off-by: Samuel Holland <samuel@sholland.org>