platform: Rename HAVE_I2C to HAVE_R_I2C
A23 and A33 are missing some peripherals that so far have been assumed
to exist. This motivates the addition of new HAVE_R_* Kconfig symbols.
Before adding them, ensure all of the existing symbols are consistent.

This symbol indicates that a platform has a specific R_I2C controller,
not that it has I2C in general.

Signed-off-by: Samuel Holland <samuel@sholland.org>