platform: Rename HAVE_RSB to HAVE_R_RSB
A23 and A33 are missing some peripherals that so far have been assumed
to exist. This motivates the addition of new HAVE_R_* Kconfig symbols.
Before adding them, ensure all of the existing symbols are consistent.

While having RSB in general and having R_RSB are effectively the same
feature, because R_RSB is the only RSB controller, the more specific
name fits in with the pattern of other symbols.

Signed-off-by: Samuel Holland <samuel@sholland.org>