Makefile: Set stack permissions for assembly objects
BFD ld expects all input objects to have a .note.GNU-stack section, or
else it emits a warning:

  warning: build/scp/arch/or1k/start.o: missing .note.GNU-stack section implies executable stack

Since we treat linker warnings as fatal, this causes a build failure.

GCC emits the relevant .section directive when it generates assembly
from C. However, standalone assembly source files do not contain this
directive, so those objects were missing the section.

Tell the assembler that we want non-executable stacks by default, which
causes it to emit the right .note.GNU-stack section without needing to
modify each assembly source file.

Note that the stack permissions in the ELF are ultimately meaningless,
because the firmware is converted to a raw binary, and the hardware has
no memory protection capability anyway.

Closes: #214
Signed-off-by: Samuel Holland <samuel@sholland.org>